Deprecate browser-worker-utils
Remove duplicated code and add tests for base64ToArrayBuffer since we had a regression with that function

resolves #1450
