--- libs/shared/browser-worker-utils/README.md
@@ -1,3 +1,0 @@
-# shared-browser-worker-utils
-
-This library was generated with [Nx](https://nx.dev).
