--- libs/shared/browser-worker-utils/project.json
@@ -1,20 +1,0 @@
-{
-  "name": "shared-browser-worker-utils",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/shared/browser-worker-utils/src",
-  "projectType": "library",
-  "tags": ["scope:worker", "scope:browser"],
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"]
-    },
-    "test": {
-      "executor": "@nx/vite:test",
-      "outputs": ["{options.reportsDirectory}"],
-      "options": {
-        "reportsDirectory": "../../../coverage/libs/shared/browser-worker-utils"
-      }
-    }
-  }
-}
