--- libs/shared/browser-worker-utils/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/browser-worker-utils';
