--- libs/shared/browser-worker-utils/vite.config.ts
@@ -1,22 +1,0 @@
-/// <reference types='vitest' />
-import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
-import { defineConfig } from 'vite';
-
-export default defineConfig(() => ({
-  root: __dirname,
-  cacheDir: '../../../node_modules/.vite/libs/shared/browser-worker-utils',
-  plugins: [nxViteTsPaths()],
-  test: {
-    name: 'shared-browser-worker-utils',
-    watch: false,
-    globals: true,
-    environment: 'jsdom',
-    include: ['{src,tests}/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
-    reporters: ['default'],
-    passWithNoTests: true,
-    coverage: {
-      reportsDirectory: '../../../coverage/libs/shared/browser-worker-utils',
-      provider: 'v8' as const,
-    },
-  },
-}));
