+++ libs/shared/utils/src/lib/__tests__/utils.spec.ts
@@ -2,6 +2,7 @@ import { DATE_FORMATS } from '@jetstream/shared/constants';
 import { QueryResults, QueryResultsColumn, SalesforceRecord } from '@jetstream/types';
 import {
   alwaysResolve,
+  base64ToArrayBuffer,
   buildDateFromString,
   dateFromTimestamp,
   flattenRecord,
@@ -1162,5 +1163,34 @@ describe('utils.getSObjectNameFromAttributes', () => {
   });
   it('returns empty string if no match', () => {
     expect(getSObjectNameFromAttributes({})).toBe('');
+  });
+});
+
+describe('utils.base64ToArrayBuffer', () => {
+  it('should convert base64 string to ArrayBuffer and decode to UTF-8 string', () => {
+    const base64 =
+      'PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiPz48UGFja2FnZSB4bWxucz0iaHR0cDovL3NvYXAuc2ZvcmNlLmNvbS8yMDA2LzA0L21ldGFkYXRhIj48dHlwZXM+PG1lbWJlcnM+QV9DdXN0b21fT2JqX19jLmFfMV90ZXN0XzFfX2M8L21lbWJlcnM+PG5hbWU+Q3VzdG9tRmllbGQ8L25hbWU+PC90eXBlcz48dmVyc2lvbj42NC4wPC92ZXJzaW9uPjwvUGFja2FnZT4=';
+    const expectedString =
+      '<?xml version="1.0" encoding="UTF-8"?><Package xmlns="http://soap.sforce.com/2006/04/metadata"><types><members>A_Custom_Obj__c.a_1_test_1__c</members><name>CustomField</name></types><version>64.0</version></Package>';
+
+    const arrayBuffer = base64ToArrayBuffer(base64);
+    const decoder = new TextDecoder('utf-8');
+    const decodedString = decoder.decode(arrayBuffer);
+
+    expect(decodedString).toEqual(expectedString);
+  });
+
+  it('should convert an empty base64 string to an empty ArrayBuffer', () => {
+    const base64 = '';
+    const arrayBuffer = base64ToArrayBuffer(base64);
+    expect(arrayBuffer.byteLength).toEqual(0);
+  });
+
+  it('should correctly convert simple ASCII base64 string', () => {
+    const base64 = btoa('Hello, World!');
+    const arrayBuffer = base64ToArrayBuffer(base64);
+    const decoder = new TextDecoder('utf-8');
+    const decodedString = decoder.decode(arrayBuffer);
+    expect(decodedString).toEqual('Hello, World!');
   });
 });
