+++ tsconfig.base.json
@@ -48,7 +48,6 @@
       "@jetstream/prisma": ["libs/prisma/src/index.ts"],
       "@jetstream/record-form": ["libs/shared/ui-record-form/src/index.ts"],
       "@jetstream/salesforce-api": ["libs/salesforce-api/src/index.ts"],
-      "@jetstream/shared/browser-worker-utils": ["libs/shared/browser-worker-utils/src/index.ts"],
       "@jetstream/shared/client-logger": ["libs/shared/client-logger/src/index.ts"],
       "@jetstream/shared/constants": ["libs/shared/constants/src/index.ts"],
       "@jetstream/shared/data": ["libs/shared/data/src/index.ts"],
